requirements.txt
pillow==12.0.0
python-dotenv==1.2.0

